school_management/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-bootstrap-form==3.4
django-bootstrap-icons==0.8.2
django-filter==22.1
idna==3.4
mysqlclient==2.1.1
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
